[Refactor] AccessToken 갱신, 자동로그인 관련 로직 수정
