+++ app/src/main/java/com/example/numberoneproject/presentation/view/MainActivity.kt
@@ -5,7 +5,6 @@ import android.os.Bundle
 import android.util.Log
 import android.view.View
 import androidx.activity.viewModels
-import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.ui.setupWithNavController
 import com.example.numberoneproject.R
@@ -14,10 +13,11 @@ import com.example.numberoneproject.databinding.ActivityMainBinding
 import com.example.numberoneproject.presentation.base.BaseActivity
 import com.example.numberoneproject.presentation.util.Extensions.repeatOnStarted
 import com.example.numberoneproject.presentation.util.TokenManager
+import com.example.numberoneproject.presentation.view.login.LoginActivity
 import com.example.numberoneproject.presentation.viewmodel.LoginViewModel
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.collectLatest
-import kotlinx.coroutines.launch
+import kotlinx.coroutines.flow.first
 import javax.inject.Inject
 
 @AndroidEntryPoint
@@ -37,6 +37,15 @@ class MainActivity : BaseActivity<ActivityMainBinding>(R.layout.activity_main) {
     }
 
     override fun subscribeUi() {
+        /** 유저 토큰정보 확인 용도 **/
+        repeatOnStarted {
+            val accessToken = tokenManager.accessToken.first()
+            val refreshToken = tokenManager.refreshToken.first()
+
+            Log.d("taag AccessToken", accessToken)
+            Log.d("taag RefreshToken", refreshToken)
+        }
+
         repeatOnStarted {
             loginVM.loginErrorState.collectLatest { response ->
                 response?.let {
@@ -45,14 +54,14 @@ class MainActivity : BaseActivity<ActivityMainBinding>(R.layout.activity_main) {
                             when(it.code) {
                                 403 -> {    // AccessToken이 만료된 경우
                                     loginVM.refreshAccessToken()
-                                    showToast("403에러 MainActivity AccessToken, RefreshToken 갱신했음")
+                                    showToast("403에러 : MainActivity에서 Token 갱신")
                                 }
                                 404 -> {    // RefreshToken이 만료된 경우
                                     showToast("MainActivity 404 에러 펑")
                                     tokenManager.writeLoginTokens("", "")
 
-                                    startActivity(Intent(this@MainActivity, LoginActivity::class.java))
                                     finish()
+                                    startActivity(Intent(this@MainActivity, LoginActivity::class.java))
                                 }
                                 else -> { showToast("MainActivity ${it.code}번 에러 펑") }
                             }
